CC/ML API/requirements.txt
fastapi
uvicorn[standard]
google-cloud-firestore
tensorflow
numpy
joblib
scikit-learn